Re-org files and simplify imports
