+++ src/helpers/subgraph/Constants.ts
@@ -1,0 +1,5 @@
+export const CATEGORY_STABLE = "Stable";
+export const CATEGORY_VOLATILE = "Volatile";
+export const CATEGORY_POL = "Protocol-Owned Liquidity";
+
+export const TOKEN_SUPPLY_TYPE_LIQUIDITY = "Liquidity";
