+++ src/views/TreasuryDashboard/components/Graph/Constants.ts
@@ -12,12 +12,6 @@ export const PARAM_TOKEN_GOHM = "gOHM";
 
 export const PARAM_SUBGRAPH = "subgraphId";
 
-export const CATEGORY_STABLE = "Stable";
-export const CATEGORY_VOLATILE = "Volatile";
-export const CATEGORY_POL = "Protocol-Owned Liquidity";
-
-export const TOKEN_SUPPLY_TYPE_LIQUIDITY = "Liquidity";
-
 export type GraphProps = {
   subgraphUrl: string;
   /**
