+++ src/views/TreasuryDashboard/components/Graph/TreasuryAssetsTable.tsx
@@ -10,7 +10,7 @@ import {
 } from "src/generated/graphql";
 import { formatCurrency } from "src/helpers";
 import { adjustDateByDays, getISO8601String } from "src/helpers/DateHelper";
-import { renameToken } from "src/helpers/ProtocolMetricsHelper";
+import { renameToken } from "src/helpers/subgraph/ProtocolMetricsHelper";
 import { ChartCard } from "src/views/TreasuryDashboard/components/Graph/ChartCard";
 import {
   AssetsTableProps,
