--- src/views/TreasuryDashboard/components/Graph/helpers/TreasuryQueryHelper.ts
@@ -1,14 +1,0 @@
-import { TokenSupply } from "src/generated/graphql";
-import {
-  getGOhmSyntheticSupply,
-  getOhmFloatingSupply,
-} from "src/views/TreasuryDashboard/components/Graph/helpers/TokenSupplyQueryHelper";
-
-export const getLiquidBackingPerOhmFloating = (liquidBacking: number, tokenSupplies: TokenSupply[]) =>
-  liquidBacking / getOhmFloatingSupply(tokenSupplies);
-
-export const getLiquidBackingPerGOhmSynthetic = (
-  liquidBacking: number,
-  currentIndex: number,
-  tokenSupplies: TokenSupply[],
-) => liquidBacking / getGOhmSyntheticSupply(currentIndex, getOhmFloatingSupply(tokenSupplies));
